deleted swing components, about window

+++ src/main/java/org/artorg/tools/phantomData/client/util/FxUtil.java
@@ -8,13 +8,15 @@ import java.util.concurrent.Executors;
 import java.util.function.Consumer;
 import java.util.function.Function;
 
+import org.artorg.tools.phantomData.client.Main;
+
 import javafx.concurrent.Task;
 import javafx.embed.swing.SwingFXUtils;
 import javafx.event.ActionEvent;
 import javafx.event.EventHandler;
 import javafx.geometry.Pos;
 import javafx.geometry.Rectangle2D;
-import javafx.scene.Node;
+import javafx.scene.Parent;
 import javafx.scene.Scene;
 import javafx.scene.control.ComboBox;
 import javafx.scene.control.ContextMenu;
@@ -25,29 +27,31 @@ import javafx.scene.control.MenuItem;
 import javafx.scene.control.TableCell;
 import javafx.scene.control.TableColumn;
 import javafx.scene.image.Image;
-import javafx.scene.layout.AnchorPane;
+import javafx.stage.Modality;
 import javafx.stage.Stage;
+import javafx.stage.StageStyle;
 import javafx.util.Callback;
 
 public class FxUtil extends org.artorg.tools.phantomData.server.util.FxUtil {
-	
-	public static void openFrame(String title, Node node, Rectangle2D bounds) {
-    	Stage stage = new Stage();
-		AnchorPane root = new AnchorPane();
-		Scene scene = new Scene(root);
-		
-		FxUtil.addToPane(root, node);
-		scene.setRoot(root);
-		
-		
+
+	public static void openFrame(String title, Parent parent) {
+		Stage stage = new Stage();
+		Scene scene = new Scene(parent);
+		scene.setRoot(parent);
 		stage.setScene(scene);
+		stage.initStyle(StageStyle.UTILITY);
+		stage.initModality(Modality.APPLICATION_MODAL);
 		stage.sizeToScene();
+		stage.setResizable(false);
 		stage.setTitle(title);
+		FxUtil.addMouseMovingSupport(stage, parent);
+		Rectangle2D bounds = new Rectangle2D(Main.getStage().getX(), Main.getStage().getY(),
+				Main.getStage().getWidth(), Main.getStage().getHeight());
 		stage.show();
+		stage.setX(bounds.getMinX() + bounds.getWidth() / 2.0 - stage.getWidth() / 2.0);
+		stage.setY(bounds.getMinY() + bounds.getHeight() / 2.0 - stage.getHeight() / 2.0);
 		
-		stage.setX(bounds.getMinX() +bounds.getWidth()/2.0 - stage.getWidth()/2.0);
-		stage.setY(bounds.getMinY() +bounds.getHeight()/2.0 - stage.getHeight()/2.0);
-    }
+	}
 
 	public static Image getFileIcon(File file) {
 		final String ext = IOutil.getFileExt(file.getPath());
@@ -64,25 +68,24 @@ public class FxUtil extends org.artorg.tools.phantomData.server.util.FxUtil {
 			} catch (IOException e) {
 				e.printStackTrace();
 			} finally {
-				if (tempFile != null)
-					tempFile.delete();
+				if (tempFile != null) tempFile.delete();
 			}
 		}
 
-		if (jswingIcon != null)
-			fileIcon = jswingIconToImage(jswingIcon);
+		if (jswingIcon != null) fileIcon = jswingIconToImage(jswingIcon);
 
 		return fileIcon;
 	}
 
 	public static Image jswingIconToImage(javax.swing.Icon jswingIcon) {
-		BufferedImage bufferedImage = new BufferedImage(jswingIcon.getIconWidth(), jswingIcon.getIconHeight(),
-				BufferedImage.TYPE_INT_ARGB);
+		BufferedImage bufferedImage = new BufferedImage(jswingIcon.getIconWidth(),
+				jswingIcon.getIconHeight(), BufferedImage.TYPE_INT_ARGB);
 		jswingIcon.paintIcon(null, bufferedImage.getGraphics(), 0, 0);
 		return SwingFXUtils.toFXImage(bufferedImage, null);
 	}
 
-	public static <T> TableColumn<T, Void> createButtonCellColumn(String text, Consumer<T> consumer) {
+	public static <T> TableColumn<T, Void> createButtonCellColumn(String text,
+			Consumer<T> consumer) {
 		TableColumn<T, Void> column = new TableColumn<T, Void>();
 		column.setCellFactory(new Callback<TableColumn<T, Void>, TableCell<T, Void>>() {
 			@Override
@@ -98,8 +101,8 @@ public class FxUtil extends org.artorg.tools.phantomData.server.util.FxUtil {
 							cell.setText("");
 							cell.setOnMouseClicked(null);
 						} else {
-							cell.setOnMouseClicked(
-									event -> consumer.accept(getTableView().getItems().get(cell.getIndex())));
+							cell.setOnMouseClicked(event -> consumer
+									.accept(getTableView().getItems().get(cell.getIndex())));
 							cell.setAlignment(Pos.CENTER);
 							cell.setText(text);
 						}
@@ -115,14 +118,16 @@ public class FxUtil extends org.artorg.tools.phantomData.server.util.FxUtil {
 
 		return column;
 	}
-	
-	public static <T> void setComboBoxCellFactory(ComboBox<T> comboBox, Function<T,String> mapper) {
+
+	public static <T> void setComboBoxCellFactory(ComboBox<T> comboBox,
+			Function<T, String> mapper) {
 		Callback<ListView<T>, ListCell<T>> cellFactory = FxUtil.createComboBoxCellFactory(mapper);
 		comboBox.setButtonCell(cellFactory.call(null));
 		comboBox.setCellFactory(cellFactory);
 	}
 
-	private static <T> Callback<ListView<T>, ListCell<T>> createComboBoxCellFactory(Function<T, String> mapper) {
+	private static <T> Callback<ListView<T>, ListCell<T>>
+			createComboBoxCellFactory(Function<T, String> mapper) {
 		return param -> {
 			return new ListCell<T>() {
 				@Override
@@ -144,7 +149,8 @@ public class FxUtil extends org.artorg.tools.phantomData.server.util.FxUtil {
 		menu.getItems().add(menuItem);
 	}
 
-	public static void addMenuItem(ContextMenu rowMenu, String name, EventHandler<ActionEvent> eventHandler) {
+	public static void addMenuItem(ContextMenu rowMenu, String name,
+			EventHandler<ActionEvent> eventHandler) {
 		MenuItem menuItem = new MenuItem(name);
 		menuItem.setOnAction(eventHandler);
 		rowMenu.getItems().add(menuItem);
